requirements.txt
git+https://github.com/Pycord-Development/pycord
# py-cord==2.0.0b7
pillow~=9.1.0
emoji~=1.7.0
numpy~=1.22.3
regex~=2022.3.15
aiohttp~=3.8.1
asyncpraw~=7.5.0
tweepy~=4.6.0
webpreview~=1.6.0
lxml~=4.8.0
steam~=1.2.1
dota2~=1.1.0
twitchapi~=2.5.3
python-dotenv~=0.20.0
vdf~=3.3
sqlalchemy~=1.4.35
psycopg2~=2.9.3
pyot~=5.3.1
git+https://github.com/meraki-analytics/role-identification.git
pytimeparse~=1.1.8
beautifulsoup4~=4.11.1
urllib3~=1.26.9
wordcloud~=1.8.1
jishaku~=2.4.0
matplotlib~=3.5.0
PyGithub~=1.55
gtts~=2.2.4
ffmpeg
requests~=2.27.1
unidiff~=0.7.3
psutil~=5.9.0